Cleaning Commands

+++ src/commands/devs/eval.js
@@ -1,6 +1,6 @@
-/* eslint-disable no-eval */
-const { inspect } = require('util')
 const { Command, Embed, Loggers } = require('../../')
+const { inspect } = require('util')
+const value = (l, c) => `\`\`\`${l}\n${String(c).slice(0, 1000) + (c.length >= 1000 ? '...' : '')}\n\`\`\``.replace(process.env.BOT_TOKEN, () => '*'.repeat(process.env.BOT_TOKEN.length))
 
 class Eval extends Command {
   constructor (client) {
@@ -11,33 +11,45 @@ class Eval extends Command {
   }
 
   async run ({ author, guild, channel, member, language, command, prefix, message, query, send, args, t, emoji }) {
-    const code = query.replace(/^```(js|javascript ?\n)?|```$/g, '')
-    const value = (l, c) => `\`\`\`${l}\n${String(c).slice(0, 1000) + (c.length >= 1000 ? '...' : '')}\n\`\`\``.replace(process.env.BOT_TOKEN, () => '*'.repeat(process.env.BOT_TOKEN.length))
     const embed = new Embed()
+
+    const code = query.replace(/^```(js|javascript ?\n)?|```$/g, '')
+
     try {
-      const resultEval = eval(code)
-      const toEval = typeof resultEval === 'string' ? resultEval : inspect(resultEval, { depth: 1 })
-      embed.addField('Result', value('js', toEval))
-      embed.addField('Type', value('css', typeof resultEval))
+      const evald = eval(code)
+      const toEval = typeof evald === 'string' ? evald : inspect(evald, { depth: 1 })
+
+      embed
+        .addField('Result', value('js', toEval))
+        .addField('Type', value('css', typeof evald))
+
       Loggers.warn(['COMMAND', 'EVAL', 'RESULT'], toEval)
     } catch (error) {
-      embed.addField('Error', value('js', error))
+      embed
+        .addField('Error', value('js', error))
+
       Loggers.error(['COMMAND', 'EVAL', 'RESULT', 'ERROR'], error)
     } finally {
       const msg = await send(embed)
-      await msg.react(emoji('CANCEL', { id: true }))
 
-      const filter = (r, u) => r.me && message.author.id === u.id
-      const collector = await msg.createReactionCollector(filter, { max: 1, errors: ['time'], time: 15000 })
+      const perms = channel.permissionsFor(guild.id)
 
-      collector.on('collect', async () => {
-        await msg.delete()
-        await message.delete()
-      })
-      collector.on('end', () => {
-        if (msg) msg.reactions.removeAll()
-      })
+      if (perms.has('ADD_REACTION') && perms.has('MANAGE_MESSAGES')) {
+        await msg.react(emoji('CANCEL', { id: true }))
+
+        const filter = (r, u) => r.me && message.author.id === u.id
+        const collector = await msg.createReactionCollector(filter, { max: 1, errors: ['time'], time: 15000 })
+
+        collector.on('collect', async () => {
+          await msg.delete()
+          await message.delete().catch()
+        })
+        collector.on('end', () => {
+          if (msg) msg.reactions.removeAll().catch(() => {})
+        })
+      }
     }
   }
 }
+
 module.exports = Eval
